Update page-icon when opening a new tab
We need to add listener to listen to onUpdated event. The
reason is that when a new tab is created, the page-icon
is the default icon which may not reflect the visibility
property.

On that other hand, this event can fire for many other reasons.
So, right now, we are setting page-icon multiple times.

To remove the redundancy, we can track if we already updated
the page-icon or not. But for the time being, this solution should
work fine.